third_party/blink/web_tests/platform/generic/external/wpt/css/selectors/has-argument-with-explicit-scope.tentative-expected.txt

This is a testharness.js-based test.
PASS :has(:scope) matches expected elements on scope1
PASS :has(:scope .c) matches expected elements on scope1
PASS :has(.a :scope) matches expected elements on scope1
PASS .a:has(:scope) .c matches expected elements on scope1
PASS .a:has(:scope) .c and :is(.a :scope .c) returns same elements on scope1
PASS .a:has(:scope) .c matches expected elements on scope2
PASS .a:has(:scope) .c and :is(.a :scope .c) returns same elements on scope2
FAIL .c:has(:is(:scope .d)) matches expected elements on scope1 assert_equals: expected "d02,d03" but got ""
FAIL .c:has(:is(:scope .d)) and :scope .c:has(.d) returns same elements on scope1 assert_equals: expected "d02,d03" but got ""
FAIL .c:has(:is(:scope .d)) and .c:has(.d) returns same elements on scope1 assert_equals: expected "d02,d03" but got ""
PASS .c:has(:is(:scope .d)) matches expected elements on scope2
PASS .c:has(:is(:scope .d)) and :scope .c:has(.d) returns same elements on scope2
PASS .c:has(:is(:scope .d)) and .c:has(.d) returns same elements on scope2
Harness: the test ran to completion.
